experiments/requirements.txt
anthropic>=0.40.0
openai>=1.50.0
httpx>=0.27.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
azure-search-documents>=11.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.0
